requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
argcomplete==1.9.3        # via zappa
base58==1.0.0             # via zappa
boto3==1.7.54
botocore==1.10.54         # via boto3, s3transfer, zappa
certifi==2018.4.16        # via requests
cfn-flip==1.0.3           # via troposphere
chardet==3.0.4            # via requests
click==6.7                # via cfn-flip, flask, kappa
docutils==0.14            # via botocore, zappa
durationpy==0.5           # via zappa
flask==1.0.2
future==0.16.0            # via zappa
hjson==3.0.1              # via zappa
idna==2.7                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
jmespath==0.9.3           # via boto3, botocore, zappa
kappa==0.6.0              # via zappa
lambda-packages==0.20.0   # via zappa
markupsafe==1.0           # via jinja2
placebo==0.8.1            # via kappa
python-dateutil==2.6.1    # via botocore, zappa
python-dotenv==0.8.2
python-slugify==1.2.4     # via zappa
pyyaml==3.12              # via cfn-flip, kappa, zappa
requests==2.19.1
s3transfer==0.1.13        # via boto3
six==1.11.0               # via cfn-flip, python-dateutil, zappa
toml==0.9.4               # via zappa
tqdm==4.19.1              # via zappa
troposphere==2.3.1        # via zappa
unidecode==1.0.22         # via python-slugify
urllib3==1.23             # via requests
werkzeug==0.14.1          # via flask, zappa
wheel==0.31.1             # via zappa
wsgi-request-logger==0.4.6  # via zappa
zappa==0.46.1
